slabtop: move unit tests to parse.rs
